# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/example-crud/auth-test": {
      "get": {
        "operationId": "ExampleCrudController_authTest",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTestDto"
                }
              }
            }
          }
        },
        "tags": [
          "Example Crud"
        ]
      }
    },
    "/example-crud/hello-world": {
      "get": {
        "operationId": "ExampleCrudController_helloWorld",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTestDto"
                }
              }
            }
          }
        },
        "tags": [
          "Example Crud"
        ]
      }
    },
    "/example-crud": {
      "post": {
        "operationId": "ExampleCrudController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExampleMongodbDocDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleMongodbDocDto"
                }
              }
            }
          }
        },
        "tags": [
          "Example Crud"
        ]
      },
      "get": {
        "operationId": "ExampleCrudController_findAll",
        "parameters": [
          {
            "name": "archived",
            "required": false,
            "in": "query",
            "description": "Filter by archived status",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExampleMongodbDocDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Example Crud"
        ]
      }
    },
    "/example-crud/{id}": {
      "get": {
        "operationId": "ExampleCrudController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleMongodbDocDto"
                }
              }
            }
          }
        },
        "tags": [
          "Example Crud"
        ]
      },
      "patch": {
        "operationId": "ExampleCrudController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExampleMongodbDocDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleMongodbDocDto"
                }
              }
            }
          }
        },
        "tags": [
          "Example Crud"
        ]
      },
      "delete": {
        "operationId": "ExampleCrudController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Example Crud"
        ]
      }
    },
    "/example-crud/name/{name}": {
      "get": {
        "operationId": "ExampleCrudController_findByName",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleMongodbDocDto"
                }
              }
            }
          }
        },
        "tags": [
          "Example Crud"
        ]
      }
    },
    "/example-crud/{id}/archive": {
      "patch": {
        "operationId": "ExampleCrudController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleMongodbDocDto"
                }
              }
            }
          }
        },
        "tags": [
          "Example Crud"
        ]
      }
    },
    "/example-crud/{id}/unarchive": {
      "patch": {
        "operationId": "ExampleCrudController_unarchive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExampleMongodbDocDto"
                }
              }
            }
          }
        },
        "tags": [
          "Example Crud"
        ]
      }
    },
    "/backstage/services": {
      "get": {
        "operationId": "BackstageCatalogController_list",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search by name, description, or topics",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "description": "Comma-separated list of blobs to include (readme,openapi,runbook,metadata)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh",
            "required": false,
            "in": "query",
            "description": "Trigger a refresh for the returned items",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceSummaryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List cached backstage services",
        "tags": [
          "Backstage Catalog"
        ]
      }
    },
    "/backstage/services/{repo}": {
      "get": {
        "operationId": "BackstageCatalogController_getOne",
        "parameters": [
          {
            "name": "repo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh",
            "required": false,
            "in": "query",
            "description": "Trigger a refresh before returning details",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get full detail for a specific repo",
        "tags": [
          "Backstage Catalog"
        ]
      }
    },
    "/backstage/sync": {
      "post": {
        "operationId": "BackstageCatalogController_sync",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Backstage Catalog"
        ]
      }
    },
    "/backstage/services/{repo}/readme": {
      "get": {
        "operationId": "BackstageCatalogController_readme",
        "parameters": [
          {
            "name": "repo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Backstage Catalog"
        ]
      }
    },
    "/backstage/services/{repo}/openapi": {
      "get": {
        "operationId": "BackstageCatalogController_openapi",
        "parameters": [
          {
            "name": "repo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Backstage Catalog"
        ]
      }
    },
    "/backstage/services/{repo}/runbook": {
      "get": {
        "operationId": "BackstageCatalogController_runbook",
        "parameters": [
          {
            "name": "repo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Backstage Catalog"
        ]
      }
    },
    "/backstage/services/{repo}/metadata": {
      "get": {
        "operationId": "BackstageCatalogController_metadata",
        "parameters": [
          {
            "name": "repo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Backstage Catalog"
        ]
      }
    }
  },
  "info": {
    "title": "NGX backstage API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "AuthTestDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "ExampleMongodbDocObjectDto": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zip": {
            "type": "string"
          }
        },
        "required": [
          "street",
          "city",
          "state",
          "zip"
        ]
      },
      "CreateExampleMongodbDocDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "SOME_ENUM",
              "SOME_OTHER_ENUM"
            ]
          },
          "description": {
            "type": "string"
          },
          "archived": {
            "type": "boolean"
          },
          "exampleMongodbDocObject": {
            "$ref": "#/components/schemas/ExampleMongodbDocObjectDto"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "ExampleMongodbDocDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "SOME_ENUM",
              "SOME_OTHER_ENUM"
            ]
          },
          "version": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          },
          "archived": {
            "type": "boolean"
          },
          "exampleMongodbDocObject": {
            "$ref": "#/components/schemas/ExampleMongodbDocObjectDto"
          },
          "__v": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "name",
          "type",
          "version",
          "description",
          "lastUpdated",
          "archived",
          "__v"
        ]
      },
      "UpdateExampleMongodbDocDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "SOME_ENUM",
              "SOME_OTHER_ENUM"
            ]
          },
          "description": {
            "type": "string"
          },
          "archived": {
            "type": "boolean"
          },
          "exampleMongodbDocObject": {
            "$ref": "#/components/schemas/ExampleMongodbDocObjectDto"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DocBlobDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "Raw content of the document"
          },
          "sha": {
            "type": "string",
            "description": "Git SHA of the fetched file"
          },
          "fetchedAt": {
            "type": "string",
            "format": "date-time"
          },
          "format": {
            "type": "string",
            "enum": [
              "yaml",
              "json",
              "markdown"
            ]
          }
        },
        "required": [
          "content"
        ]
      },
      "ServiceSummaryDto": {
        "type": "object",
        "properties": {
          "repoName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "topics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "defaultBranch": {
            "type": "string"
          },
          "htmlUrl": {
            "type": "string"
          },
          "lastSyncAt": {
            "type": "string",
            "format": "date-time"
          },
          "syncStatus": {
            "type": "string",
            "enum": [
              "idle",
              "ok",
              "partial",
              "failed",
              "rate_limited"
            ]
          },
          "syncError": {
            "type": "string"
          },
          "readme": {
            "$ref": "#/components/schemas/DocBlobDto"
          },
          "openapi": {
            "$ref": "#/components/schemas/DocBlobDto"
          },
          "runbook": {
            "$ref": "#/components/schemas/DocBlobDto"
          },
          "serviceMetadata": {
            "$ref": "#/components/schemas/DocBlobDto"
          }
        },
        "required": [
          "repoName",
          "topics",
          "syncStatus"
        ]
      },
      "ServiceDetailDto": {
        "type": "object",
        "properties": {
          "repoName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "topics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "defaultBranch": {
            "type": "string"
          },
          "htmlUrl": {
            "type": "string"
          },
          "lastSyncAt": {
            "type": "string",
            "format": "date-time"
          },
          "syncStatus": {
            "type": "string",
            "enum": [
              "idle",
              "ok",
              "partial",
              "failed",
              "rate_limited"
            ]
          },
          "syncError": {
            "type": "string"
          },
          "readme": {
            "$ref": "#/components/schemas/DocBlobDto"
          },
          "openapi": {
            "$ref": "#/components/schemas/DocBlobDto"
          },
          "runbook": {
            "$ref": "#/components/schemas/DocBlobDto"
          },
          "serviceMetadata": {
            "$ref": "#/components/schemas/DocBlobDto"
          }
        },
        "required": [
          "repoName",
          "topics",
          "syncStatus"
        ]
      },
      "SyncRequestDto": {
        "type": "object",
        "properties": {
          "repos": {
            "description": "Optional list of repo names to sync; defaults to all repos in the org",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "force": {
            "type": "boolean",
            "description": "Force re-fetch even if SHA is unchanged"
          }
        }
      },
      "SyncResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "succeeded": {
            "type": "number"
          },
          "failed": {
            "type": "number"
          },
          "durationMs": {
            "type": "number",
            "description": "Duration in milliseconds"
          },
          "failures": {
            "description": "Repos that failed to sync (with error message inline if available)",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "total",
          "succeeded",
          "failed",
          "durationMs",
          "failures"
        ]
      }
    }
  }
}